automatically enable egui-winit `clipboard` feature
This chooses a different tradeoff where it is more easily possible to get rid
of x11 or wayland dependencies when they are not needed. Otherwise, when you
want `clipboard` and `x11` you will still pull in lots of wayland dependencies
for wayland clipboard support that will not be used anyway.