Fix issue #2578 - `row_start_x` tracks a virtual position in the source paragraph (the one that is too long) for which length has already been processed. When creating an empty row, this position should not be updated as no glyphs were consumed from the source paragraph. - added example that would demonstrate the problem if the line was included, and that is fixed with this commit
Fix issue #2578 There was confusion in the code over how to
break when on a non-empty visual row (`first_row_indentation > 0.0`), causing
text to be shifted left outside the ui frame. This is the case for example
when another label has already been placed in this `ui.horizontal_wrapped()`.
This fix will not create an empty row, essentially starting a newline, but
rather try to fit as much text as possible on the existing row. IMO this is
the desired use of a wrapping layout. I've also added an example that would
demonstrate the problem if the line was included, and that is fixed with this
commit