Fix visual glitch on the right side of highly rounded rectangles
When one side of a rectangle is all rounding we need to take care
not to produce duplicated vertices in the rectangle path generator.

The old code only handled three sides, but forgot the last side
(the right side).
The new code handles the right side, and also handles the other
sides more robustly (with a floating point eps) and efficiently
(in a single pass).