Add a `Display` impl for `Vec2`, `Pos2`, and `Rect`
These three types currently have a `Debug` implementation that only
ever prints one decimal point. Sometimes it is useful to see more of the
number, or otherwise have specific formatting.

Add `Display` implementations that pass the format specification to the
member `f32`s for an easier way to control what is shown when debugging.