Decouple directories from persistence in eframe
It is now possible to enable the persistence feature without depending
on the directories crate, for uses who wish to provide their own paths
from another source.

When the directories feature is not enabled, a simpler naive
approximation of its output is used, which covers the most common target
systems.